refactor: InjectionKey -> media.types.ts moved
